+++ plugins/scaffolder-annotator-action/src/index.ts
@@ -1,0 +1,9 @@
+/***/
+/**
+ * The annotator module for @backstage/plugin-scaffolder-backend.
+ *
+ * @packageDocumentation
+ */
+
+export * from './actions';
+export * from './dynamic';